Se limpia codigo y se agrega nuevo controller para dividir las responsabilidades.
