+++ Resources/SaveUserResource.cs
@@ -4,8 +4,6 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
-using System.ComponentModel.DataAnnotations;
-
 
 namespace GoingTo_API.Resources
 {
