+++ Services/CountryService.cs
@@ -14,15 +14,11 @@ namespace GoingTo_API.Services
     {
         private readonly ICountryRepository _countryRepository;
         private readonly ICountryCurrencyRepository _countryCurrencyRepository;
-        private readonly ICountryLanguageRepository _countryLanguageRepository;
-        private readonly IUnitOfWork _unitOfWork;
         
-        public CountryService(ICountryRepository countryRepository,IUnitOfWork unitOfWork, ICountryCurrencyRepository countryCurrencyRepository, ICountryLanguageRepository countryLanguageRepository)
+        public CountryService(ICountryRepository countryRepository,ICountryCurrencyRepository countryCurrencyRepository)
         {
             _countryRepository = countryRepository;
-            _unitOfWork = unitOfWork;
             _countryCurrencyRepository = countryCurrencyRepository;
-            _countryLanguageRepository = countryLanguageRepository;
         }
         public async Task<CountryResponse> GetByIdAsync(int id)
         {
